Add files via upload
Radio Propagation Models (Multi-wall and Log-Distance Shadowing). It also includes an optimization algorithm that finds the propagation parameters if field measurements are conducted (signal strength). Multiple transmitters can be added as well.
I have also a MATLAB code similar to this, but it is not as complete as this Python version. There is a 3D ray tracing I wrote in MATLAB as well that you can check. Please read the readme.dox  or readme.txt
If you've questions please feel free to ask